requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
chromadb==1.2.1
click==8.3.0
coloredlogs==15.0.1
dataclasses-json==0.6.7
datasets==4.3.0
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
email-validator==2.3.0
fastapi==0.119.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2023.10.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.4.1
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
instructor==1.11.3
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.0.2
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-google-genai==3.0.0
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.37
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.15
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.109.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.11.0
overrides==7.7.0
packaging==23.2
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
PyPika==0.48.9
pyproject_hooks==1.2.0
pysbd==0.3.4
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
ragas==0.3.7
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.18.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scikit-network==0.33.3
scipy==1.16.2
sentence-transformers==5.1.2
sentry-sdk==2.42.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
